control: Fix InterfacesRemoved match string
The interfaces_removed trigger was using the wrong match for the
InterfacesRemoved signal.  It was trying look for the object path in
question to be the 'path' field, but 'path' is just the path of the
object manager sending the signal.  It should be looking at argument 0
instead.  This was previously fixed for InterfacesAdded but somehow
missed for this one.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: If818b228472606dd204d507cfe04c7bacdc5fc03