monitor: Remove presence check in startMonitor() (#31)
Before this fix, there was a window where if a fan was removed after
power on but before the monitor start timer expired, no power off
actions would run.  This was because the tach sensor value was always
zero, so there would be no tach changed signals so no handlers would
run to start the countdown to the fan becoming nonfunctional.

The startMonitor() function was calling tachChanged() already manually,
but only for present fans.  Change the code to call it regardless of if
the fan is present or not, so that tachChanged() can properly handle the
tach sensor values of zero.


Change-Id: I554e297aeb3f9efb0007721bd8c6305fa706cfed

Signed-off-by: Matt Spinler <spinler@us.ibm.com>