control:config:ibm: Use count_state_floor delay
Add a 5 second delay to the count_state_floor action that watches for
non functional sensors to all IBM configurations.  With this change, a
sensor has to be nonfunctional for at least 5 seconds for the floor to
change.

This is being done because the NVMe temperature sensors would briefly be
set to nonfunctional when a drive is removed.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I6fa70f2551c11b2cfe8ee4696b6d72d57a3a7bd5