monitor:configs: Increase BlueRidge2U deviation
There is a period of time in the lab where the BlueRidge 2U will have to
support both the ship level and lab-only fan models.  To support both
with the same factor/offset values, increase the deviation by 4 so that
the lab-only fans with the slower outlet rotors won't have faults.

After the older lab-only fans have been replaced with the new fans, this
can be reverted.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I6fd6bb4bd59651983e7536bfd21d8842a8a88381